requirements.txt
playwright==1.40.0
supabase==2.3.5
python-dotenv==1.0.0
aiohttp==3.9.1
croniter==2.0.1
